chatbox/pyproject.toml
[project]
name = "chatbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.9.3",
    "mcp-python>=0.1.4",
    "path>=17.1.0",
    "anthropic>=0.53.0",
    "openai>=1.86.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.2",
    "langchain-core>=0.1.10",
    "langchain-community>=0.0.13",
    "langgraph>=0.0.23",
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "python-dotenv>=1.0.0",
]
